rename CodeExamplesBox -> CodeExampleBox
